requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
python-multipart
pydantic
pydantic-settings
email-validator

sentence-transformers
faiss-cpu
numpy

PyMuPDF
python-docx
Pillow
pytesseract
pdf2image

openai
